Move new classs and remove obsolete ones
