+++ modules/search-provision/src/main/scala/io/renku/search/provision/handler/MessageReader.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.search.provision.variant
+package io.renku.search.provision.handler
 
 import cats.effect.Sync
 import cats.syntax.all.*
