+++ modules/search-provision/src/main/scala/io/renku/search/provision/handler/QueueMessageDecoder.scala
