+++ modules/search-provision/src/main/scala/io/renku/search/provision/handler/ShowInstances.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.search.provision.variant
+package io.renku.search.provision.handler
 
 import io.renku.events.v1.*
 import cats.Show
